ICMP TooBig (#50)
MTU Too Big Handling & Test
---------------------------
* Handle Icmp TOOBIG Packets
* includes directory changes and cleanup for icmp in general (incl. test)
  - incl checksums for icmpv6 for generation
* checksumming operations work over as a trait now
* move static bytes for dpdk-unit tests into the tests module
* call example tests from makefile with env TEST var
* run cargo fmt on allthethings
* remove to_be/from_be for u8 types as that's pretty much the identity fn
as the effect is only on types beyond u/i`8`
* mtu-too-big test with assertions
* Further cleanup of Icmp modules
* cleanup of headerupdates trait to be a bit more generic and use PreviousHdr
* remove need for RUNNABLE_TESTS env variable
* kill background proc in srv6-inject test and update timing

+++ test/packet-generation/src/nf.rs
@@ -23,12 +23,13 @@ impl PacketCreator {
         };
         mac.set_etype(mac::EtherType::IPv4);
         let mut ip = Ipv4Header::new();
-        ip.set_src(u32::from(Ipv4Addr::from_str("10.0.0.1").unwrap()));
-        ip.set_dst(u32::from(Ipv4Addr::from_str("10.0.0.5").unwrap()));
+        ip.set_src(Ipv4Addr::from_str("10.0.0.1").unwrap());
+        ip.set_dst(Ipv4Addr::from_str("10.0.0.5").unwrap());
         ip.set_ttl(128);
         ip.set_version(4);
         ip.set_ihl(5);
         ip.set_length(20);
+
         PacketCreator {
             mac: mac,
             ip: ip,
